Relocate file.
