+++ src/combine.js
